Remove queries.go
