+++ source/iterator/constants.go
@@ -33,4 +33,43 @@ const (
 
 	timeoutBeforeCloseDBSec        = 20
 	timeoutToClearTrackingTableSec = 5
+
+	// queries patterns.
+	queryIfTableExists     = "SELECT table_name FROM user_tables WHERE table_name='%s'"
+	querySnapshotTable     = "CREATE SNAPSHOT %s AS SELECT * FROM %s"
+	queryTableCopy         = "CREATE TABLE %s AS SELECT * FROM %s WHERE 1=2 UNION ALL SELECT * FROM %s WHERE 1=2"
+	queryTriggerInsertPart = "INSERT INTO %s (%s, %s) VALUES (%s, transaction_type)"
+	querySelectRowsFmt     = "SELECT %s FROM %s%s ORDER BY %s ASC FETCH NEXT %d ROWS ONLY"
+	queryDeleteByIDs       = "DELETE FROM %s WHERE %s IN (%s)"
+
+	queryTrackingTableExtendWithConduitColumns = `ALTER TABLE %s 
+ADD (
+	%s NUMBER GENERATED BY DEFAULT ON NULL AS IDENTITY,
+    %s VARCHAR2(6) NOT NULL CHECK(%s IN ('insert','update','delete')),
+    %s DATE DEFAULT SYSDATE,
+	
+	CONSTRAINT %s UNIQUE (%s)
+)`
+
+	queryTriggerCreate = `
+CREATE OR REPLACE TRIGGER %s
+    AFTER
+    INSERT OR UPDATE OR DELETE
+    ON %s
+    FOR EACH ROW
+DECLARE
+   transaction_type VARCHAR2(6);
+BEGIN
+   transaction_type := CASE
+         WHEN INSERTING THEN 'insert'
+         WHEN UPDATING THEN 'update'
+         WHEN DELETING THEN 'delete'
+   END;
+
+   IF INSERTING OR UPDATING THEN
+   %s
+   ELSE
+   %s
+   END IF;
+END;`
 )
